3.0 beta update II rework
Complete rework of intern funcionality. Now uses classes Show, Season,
Episode everywhere, not just for the smartRename. Many features are not
working due to this rework. They will be back before 3.0 RC . Json is
temporarily not part of the exe and is again external dll, this is
caused by my experiments.
Several internal polishes
So much of commented code and one crazy ternary operator abomination.
All to change in future?